Print 50 first fibonacci sequence'

+++ 0x02-functions_nested_loops/101-natural.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 	int start = 0;
-	long int sum = 0;
+	int sum = 0;
 
 	while (start < 1024)
 	{
