+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,7 +7,7 @@
 int main(void)
 {
 	int count;
-	int fib1 = 1, fib2 = 2, fib3;
+	long int fib1 = 1, fib2 = 2, fib3;
 
 	printf("%lu, %lu, ", fib1, fib2);
 	count = 3;
